+++ ios/Extensions/ZelloTranscription+Extension.swift
@@ -1,0 +1,18 @@
+import ZelloSDK
+
+extension ZelloTranscription {
+  var jsonDictionary: [AnyHashable: Any] {
+    return [
+      "text": text,
+      "language": language,
+      "isTruncated": isTruncated,
+      "confidence": confidence,
+      "translations": translations.map { translation in
+        [
+          "text": translation.text,
+          "language": translation.language
+        ]
+      }
+    ]
+  }
+}